+++ Modern JavaScript From The Beginning/Section 4: DOM Projects/34 - 36. Task List/app.js
@@ -5,6 +5,18 @@ const clearButton = document.querySelector('.clear-tasks');
 const filter = document.querySelector('#filter');
 const taskInput = document.querySelector('#task');
 
+function addTaskToStorage (task) {
+    let tasks = [];
+
+    if ('tasks' in localStorage) {
+        tasks = JSON.parse(localStorage.getItem('tasks'));
+    }
+
+    tasks.push(task);
+
+    localStorage.setItem('tasks', JSON.stringify(tasks));
+}
+
 function addTask (event) {
     event.preventDefault();
 
@@ -30,16 +42,43 @@ function addTask (event) {
     // apend li to ul
     taskList.appendChild(li);
 
+    // store in local storage
+    addTaskToStorage(taskInput.value);
+
     // clear input
     taskInput.value = '';
+}
+
+function removeTaskFromStorage (taskItem) {
+    let tasks = [];
+
+    if ('tasks' in localStorage) {
+        tasks = JSON.parse(localStorage.getItem('tasks'));
+    }
+
+    tasks.forEach(
+        (task, index) => {
+            if (taskItem.textContent === task) {
+                tasks.splice(index, 1);
+            }
+        }
+    );
+
+    localStorage.setItem('tasks', JSON.stringify(tasks));
 }
 
 function deleteTask (event) {
     if (event.target.parentElement.classList.contains('delete-item')) {
         if (confirm('Are you sure?')) {
             event.target.parentElement.parentElement.remove();
+            // remove from local storage
+            removeTaskFromStorage(event.target.parentElement.parentElement);
         }
     }
+}
+
+function clearStorage () {
+    localStorage.removeItem('tasks');
 }
 
 function clearTasks () {
@@ -47,6 +86,8 @@ function clearTasks () {
         while (taskList.firstChild) {
             taskList.removeChild(taskList.firstChild);
         }
+
+        clearStorage();
     }
 }
 
@@ -66,7 +107,39 @@ function filterTasks (event) {
     );
 }
 
+function loadTasks () {
+    let tasks = [];
+
+    if ('tasks' in localStorage) {
+        tasks = JSON.parse(localStorage.getItem('tasks'));
+    }
+
+    tasks.forEach(
+        task => {
+            const li = document.createElement('li');
+            // add class
+            li.className = 'collection-item';
+            // create text node and append to li
+            li.appendChild(document.createTextNode(task));
+            // create new link element
+            const link = document.createElement('a');
+            // add class
+            link.className = 'delete-item secondary-content';
+            link.href = '#';
+            // add icon html
+            link.innerHTML = '<i class="fa fa-remove"></i>';
+            // append link to li
+            li.appendChild(link);
+            // apend li to ul
+            taskList.appendChild(li);
+        }
+    )
+}
+
 function loadEventListeners () {
+    // DOM load event
+    document.addEventListener('DOMContentLoaded', loadTasks);
+
     // add task event
     form.addEventListener('submit', addTask);