+++ CHANGELOG.md
@@ -5,8 +5,10 @@
   - Section 4: DOM Projects
     - Task List [Part 1] - UI & Add Task Items
     - Task List [Part 2] - Delete & Filter Tasks
+    - Task List [Part 3] - Persist To Local Storage
 - UI and add task logic were added.
 - Delete and filter logic was added.
+- Interaction with local storage was added.
 
 ## August 20, 2020
 - Modern JavaScript From The Beginning (Udemy)